YTA-2946: Moved setup artifacts shared by Contract and Integration tests to the 'common' package.
