+++ test/common/WSClientSetup.scala
@@ -14,15 +14,17 @@
  * limitations under the License.
  */
 
-package contract
+package common
 
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.github.fge.jsonschema.core.report.ProcessingReport
-import com.github.fge.jsonschema.main.JsonSchemaFactory
+import org.scalatest.Suite
+import org.scalatestplus.play.OneServerPerSuite
+import play.api.libs.ws.WSClient
+import uk.gov.hmrc.play.http.HeaderCarrier
 
-case class Schema(schemaPath: String) {
-  private val validator = JsonSchemaFactory.byDefault().getJsonSchema(schemaPath)
+trait WSClientSetup extends OneServerPerSuite { self: Suite =>
+  val baseUrl = s"http://localhost:$port"
 
-  def validate(json: String): ProcessingReport =
-    validator.validate(new ObjectMapper().readTree(json), true)
+  implicit val hc = HeaderCarrier()
+
+  implicit val wsClient: WSClient = app.injector.instanceOf[WSClient]
 }
